Refactored playback code in app.component.ts to individual components
I also added the framework to make it easier to add new algorithms in
the future. (abstracted relevant code to a service)

In this commit, I also added some components from Angular Material
(form elements).

Contributes to #4, #11
