+++ src/app/app-routing.module.ts
@@ -1,8 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { AlgorithmPageComponent } from './algorithm-page/algorithm-page.component';
 
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: '', component: AlgorithmPageComponent },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
